# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KanbanFlow API",
    "version": "1.0.0",
    "description": "Backend API for KanbanFlow task management application"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Subtask": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "isCompleted": { "type": "boolean" }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "tag": {
            "type": "string",
            "nullable": true,
            "enum": ["technical", "concept", "design", "blocker", "marketing", "deployment", "documentation", null]
          },
          "assignees": {
            "type": "array",
            "items": { "type": "string" }
          },
          "subtasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Subtask" }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Member": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["owner", "editor", "viewer"]
          },
          "avatar": { "type": "string", "nullable": true },
          "joinedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Column": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "tasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          }
        }
      },
      "Board": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "userId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "members": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Member" }
          },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Column" }
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/boards": {
      "get": {
        "tags": ["Boards"],
        "summary": "Get all boards",
        "description": "Returns all boards the authenticated user has access to.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Boards retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "boards": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Board" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Boards"],
        "summary": "Create a new board",
        "description": "Creates a new board with optional initial columns.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Board name"
                  },
                  "columns": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Optional list of column names"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Board created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "board": { "$ref": "#/components/schemas/Board" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}